pricing type updated, unit recalculated for confirmation and cancellation fixed with delete method

+++ messages/src/main/java/com/thundashop/zauiactivity/dto/TaxData.java
@@ -19,8 +19,8 @@ public class TaxData {
     private Integer rate;
     @SerializedName("priceExcludingTax")
 
-    private Long priceExcludingTax;
+    private double priceExcludingTax;
     @SerializedName("taxAmount")
 
-    private Long taxAmount;
+    private double taxAmount;
 }
